perf: models decimal→double e remoção de casts redundantes
Decimal tem overhead de operações de 128 bits e não é necessário para
as features do modelo (todas normalizadas [0,1]). Double é suficiente
em precisão e nativo na CPU. Remove casts explícitos (double)→float
que se tornaram desnecessários após a mudança de tipo.

+++ src/FMLab.RinhaDeBackend.DeteccaoDeFraude.Api/Models/Customer.cs
@@ -5,7 +5,7 @@ namespace FMLab.RinhaDeBackend.DeteccaoDeFraude.Models;
 public class Customer
 {
     [JsonPropertyName("avg_amount")]
-    public decimal AverageAmount { get; init; }
+    public double AverageAmount { get; init; }
     
     [JsonPropertyName("tx_count_24h")]
     public int TransactionsLast24h { get; init; }
